"free software" instead of "free application"
The sentence "LibreTaxi is free application." has two problems:
(1) incorrect grammar, since application is a countable noun, so would need to write "a" free application.
(2) the phrase "free application" is not very precise (and you already established that it was free of charge).  I believe you mean to say that it is "free software," which has a precise meaning.